+++ CHANGELOG.md
@@ -22,6 +22,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   there are no mappings in the Git fetch refspecs.
   [#5115](https://github.com/jj-vcs/jj/issues/5115)
 
+* Conflicted bookmarks and tags in `trunk()` will no longer generate verbose
+  warnings. The configured `trunk()` alias will temporarily be disabled.
+  [#8501](https://github.com/jj-vcs/jj/issues/8501)
+
 ## [0.37.0] - 2026-01-07
 
 ### Release highlights